Removed frontend folder temporarily. Resumed work on image transformer.
