+++ app/services/backend/src/image-transformer/main.py
@@ -15,12 +15,13 @@ logger.setLevel(logging.INFO)
 
 
 class ImageTransformer:
-    def __init__(self, images):
+    def __init__(self, args):
         logger.info("Starting image transforming.")
         self.input_img_path = str(
-            Path(__file__).resolve().parents[1] / "images" / images[0]
+            Path(__file__).resolve().parents[1] / "images" / args[0]
         )
-        self.output_img_name = images[1]
+        self.output_img_name = args[1]
+        self.threshold = int(args[2])
         self.image = None
         self.original_image = None
 
@@ -38,8 +39,8 @@ class ImageTransformer:
         :return: Numpy image array
         :rtype: _type_
         """
-        # Loading image
-        img = cv2.imread(self.input_img_path)
+        # Loading image (in grayscale)
+        img = cv2.imread(self.input_img_path, 0)
         if img is not None:
             logger.info("Reading image - Success.")
             self.image = img
@@ -47,13 +48,33 @@ class ImageTransformer:
         else:
             return False
 
+    def structure_edges(self):
+        logger.info("Structuring edges...")
+        # Loading model for edge detector
+        edge_detector = cv2.ximgproc.createStructuredEdgeDetection(
+            "data-models/model.yml"
+        )
+        # Converting image colors
+        src = cv2.cvtColor(self.image, cv2.COLOR_BGR2RGB)
+        # Detecting edges (pixel values have to be converted to 0 - 1)
+        edges = edge_detector.detectEdges(np.float32(src) / 255.0)
+        # Saving edge detected image to self (rescaling the pixel values)
+        self.image = 255 * edges
+        logger.info("Structuring edges - Success.")
+
     def refine_edges(self):
         """Refines edges of the image and turns the image into black & white."""
-        # Apply thresholding to image
-        img = cv2.threshold(np.uint8(self.image), 38, 255, cv2.THRESH_BINARY_INV)[1]
-        morphed = cv2.morphologyEx(img, cv2.MORPH_CLOSE, np.ones((4, 4), np.uint8))
+        logger.info("Refining edges...")
+        # Apply thresholding to image (38)
+        img = cv2.threshold(np.uint8(self.image), self.threshold,
+                            255, cv2.THRESH_BINARY_INV)[1]
+        # img = cv2.adaptiveThreshold(
+        #    np.uint8(self.image), 255, cv2.ADAPTIVE_THRESH_GAUSSIAN_C, cv2.THRESH_BINARY, 11, 2)
+        morphed = cv2.morphologyEx(
+            img, cv2.MORPH_CLOSE, np.ones((4, 4), np.uint8))
         # Finding contours
-        contours = cv2.findContours(morphed, cv2.RETR_LIST, cv2.CHAIN_APPROX_NONE)[0]
+        contours = cv2.findContours(
+            morphed, cv2.RETR_LIST, cv2.CHAIN_APPROX_NONE)[0]
         cnts = sorted(contours, key=cv2.contourArea, reverse=True)
         # Array that contains contours that will be removed
         noisy_contours = []
@@ -75,20 +96,7 @@ class ImageTransformer:
         median = cv2.medianBlur(morphed, 5)
         self.image = median
         logger.info("Refining edges - Success.")
-        cv2.imwrite("edges.png", self.image)
-
-    def structure_edges(self):
-        # Loading model for edge detector
-        edge_detector = cv2.ximgproc.createStructuredEdgeDetection(
-            "data-models/model.yml"
-        )
-        # Converting image colors
-        src = cv2.cvtColor(self.image, cv2.COLOR_BGR2RGB)
-        # Detecting edges
-        edges = edge_detector.detectEdges(np.float32(src) / 255.0)
-        # Saving edge detected image to self
-        self.image = 255 * edges
-        logger.info("Structuring edges - Success.")
+        cv2.imwrite(self.output_img_name, self.image)
 
     def perform_meanshift(self):
         """Performs meanshift operation on the target image."""
@@ -98,7 +106,8 @@ class ImageTransformer:
         # based on r g b intensities
         flattened_img = np.reshape(self.image, [-1, 3])
         # Estimate bandwidth for meanshift algorithm
-        bandwidth = estimate_bandwidth(flattened_img, quantile=0.12, n_samples=500)
+        bandwidth = estimate_bandwidth(
+            flattened_img, quantile=0.12, n_samples=500)
         ms = MeanShift(bandwidth=bandwidth, bin_seeding=True)
         # Performing meanshift on image
         ms.fit(flattened_img)
@@ -111,16 +120,17 @@ class ImageTransformer:
 
 
 def main(args):
+    logger.error(args)
     it = ImageTransformer(args)
     it.transform_image()
 
 
 if __name__ == "__main__":
     args = sys.argv
-    # Check that both input and output image names are supplied.
-    if len(args) == 3:
+    # Check that both input and output image names and edge threshold are supplied.
+    if len(args) == 4:
         main(args[1:])
     else:
         print(
-            "Supply input and output image names as arguments. Ex. python main.py input.png output.png"
+            "Supply input, output image names and edge threshold (1-50) as arguments. Ex. python main.py input.png output.png 15"
         )
