--- app/services/frontend/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "art-instruction-vue",
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "vite",
-    "build": "run-p type-check build-only",
-    "preview": "vite preview",
-    "test:unit": "vitest --environment jsdom --root src/",
-    "test:e2e": "start-server-and-test preview :4173 'cypress run --e2e'",
-    "test:e2e:dev": "start-server-and-test 'vite dev --port 4173' :4173 'cypress open --e2e'",
-    "build-only": "vite build",
-    "type-check": "vue-tsc --noEmit -p tsconfig.vitest.json --composite false",
-    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore"
-  },
-  "dependencies": {
-    "pinia": "^2.0.23",
-    "vue": "^3.2.41",
-    "vue-router": "^4.1.5"
-  },
-  "devDependencies": {
-    "@rushstack/eslint-patch": "^1.1.4",
-    "@types/jsdom": "^20.0.0",
-    "@types/node": "^16.11.68",
-    "@vitejs/plugin-vue": "^3.1.2",
-    "@vitejs/plugin-vue-jsx": "^2.0.1",
-    "@vue/eslint-config-prettier": "^7.0.0",
-    "@vue/eslint-config-typescript": "^11.0.0",
-    "@vue/test-utils": "^2.1.0",
-    "@vue/tsconfig": "^0.1.3",
-    "cypress": "^10.10.0",
-    "eslint": "^8.22.0",
-    "eslint-plugin-cypress": "^2.12.1",
-    "eslint-plugin-vue": "^9.3.0",
-    "jsdom": "^20.0.1",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.7.1",
-    "start-server-and-test": "^1.14.0",
-    "typescript": "~4.7.4",
-    "vite": "^3.1.8",
-    "vitest": "^0.24.3",
-    "vue-tsc": "^1.0.8"
-  }
-}
